add some validation fixes to delete/posts/{id}

+++ src/posts/routes/handlers/delete-post.handler.ts
@@ -1,23 +1,25 @@
 import { Request, Response } from "express";
+import { ObjectId } from "mongodb";
 import { HttpStatus } from "../../../core/types/http-statuses";
-import { createErrorMessages } from "../../../core/utils/error.utils";
 import { postsRepository } from "../../repositories/posts.repository";
 
 export async function deletePostHandler(
   req: Request<{ id: string }>,
   res: Response,
 ) {
-  const id = req.params.id;
-  const post = await postsRepository.findById(id);
   try {
-    if (!post) {
-      res
-        .status(HttpStatus.NotFound)
-        .send(
-          createErrorMessages([{ field: "id", message: "Post not found" }]),
-        );
-      return;
+    const id = req.params.id;
+
+    if (!ObjectId.isValid(id)) {
+      return res.sendStatus(HttpStatus.NotFound);
     }
+
+    const post = await postsRepository.findById(id);
+
+    if (!post) {
+      return res.sendStatus(HttpStatus.NotFound);
+    }
+
     await postsRepository.delete(id);
     res.sendStatus(HttpStatus.NoContent);
   } catch (err) {
